# Improve BMP performance by using no PNG compression during conversion